[SpellChecker] js->ts
