+++ package.json
@@ -33,6 +33,7 @@
     "@material-ui/lab": "^4.0.0-alpha.56",
     "@material-ui/styles": "^4.10.0",
     "@matt-block/react-recaptcha-v2": "^1.0.6",
+    "@types/nspell": "^2.1.0",
     "axios": "^0.19.2",
     "dictionary-en-us": "^2.1.1",
     "make-dir": "^3.1.0",
